requirements.txt
discord-py-interactions[voice]==5.14.0
python-dotenv==1.0.1
yt-dlp==2025.3.21